A first try on State.
