+++ pubspec.yaml
@@ -29,6 +29,7 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.0
   english_words: ^3.1.5
+  provider: ^4.3.2+2
 
 dev_dependencies:
   flutter_test:
